requirements.txt
alembic==1.6.5
click==8.0.1
Flask==2.1.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
gunicorn==23.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.9
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
python-dotenv==0.21.1